requirements.txt
streamlit
langchain-deepseek
langchain-community
chromadb
rank-bm25
jieba
python-dotenv